drware navigation added and header code move into a component
