+++ components/CustomDrawer.js
@@ -1,0 +1,26 @@
+import React from "react";
+import { View } from "react-native";
+import {
+    DrawerContentScrollView,
+    DrawerItemList,
+} from "@react-navigation/drawer";
+
+const CustomDrawer = (props) => {
+    return (
+        <View style={{ flex: 1 }}>
+            <DrawerContentScrollView {...props}>
+                <View
+                    style={{
+                        flex: 1,
+                        backgroundColor: "#fff",
+                        paddingTop: 10,
+                    }}
+                >
+                    <DrawerItemList {...props} />
+                </View>
+            </DrawerContentScrollView>
+        </View>
+    );
+};
+
+export default CustomDrawer;
