+++ package.json
@@ -10,13 +10,25 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@expo/vector-icons": "^13.0.0",
+    "@react-native-community/masked-view": "^0.1.11",
+    "@react-navigation/drawer": "^6.4.3",
+    "@react-navigation/native": "^6.0.11",
+    "@react-navigation/native-stack": "^6.7.0",
     "expo": "~45.0.0",
     "expo-status-bar": "~1.3.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-native": "0.68.2",
+    "react-native-gesture-handler": "~2.2.1",
     "react-native-heroicons": "^2.2.0",
+    "react-native-pager-view": "5.4.15",
+    "react-native-paper": "^4.12.2",
+    "react-native-reanimated": "~2.8.0",
+    "react-native-safe-area-context": "4.2.4",
+    "react-native-screens": "~3.11.1",
     "react-native-svg": "^12.3.0",
+    "react-native-vector-icons": "^9.2.0",
     "react-native-web": "0.17.7"
   },
   "devDependencies": {
